requirements.txt
pyecharts
pyqt
numpy